Add PGP metadata extraction and tests
Implemented a get_metadata() method in the PGP unpacker to identify PGP file types (public key, private key, encrypted message, signature) from both ASCII-armored and binary formats. Added unit tests to verify metadata extraction for various PGP data types. Improved resource cleanup in the unpack method.

+++ sflock/unpack/pgp.py
@@ -4,6 +4,7 @@
 # See the file 'docs/LICENSE.txt' for copying permission.
 
 import os
+import shutil
 import subprocess
 import tempfile
 
@@ -25,23 +26,61 @@ class PGP(Unpacker):
         else:
             filepath = self.f.temp_path()
             temporary = True
-        # ToDo
-        # locked system call occurred during sandboxing!\nip=0x7f9d2bc0fa97 sp=0x7ffdcb8d5eb8 abi=0 nr=102 syscall=getuid
-        # ret = self.zipjail(filepath, dirpath, "-o", os.path.join(dirpath, "extracted"), "--passphrase=%s" % (password or ""), filepath)
-        p = subprocess.Popen(
-            (self.exe, "--decrypt", "--batch", "-o", os.path.join(dirpath, "extracted"), "--passphrase=%s" % (password or ""), filepath),
-            stdin=subprocess.PIPE,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
 
-        return_code = p.wait()
-        _, _ = p.communicate()
-        ret = not return_code
-        if not ret:
+        try:
+            # ToDo
+            # locked system call occurred during sandboxing!\nip=0x7f9d2bc0fa97 sp=0x7ffdcb8d5eb8 abi=0 nr=102 syscall=getuid
+            # ret = self.zipjail(filepath, dirpath, "-o", os.path.join(dirpath, "extracted"), "--passphrase=%s" % (password or ""), filepath)
+            p = subprocess.Popen(
+                (self.exe, "--decrypt", "--batch", "-o", os.path.join(dirpath, "extracted"), "--passphrase=%s" % (password or ""), filepath),
+                stdin=subprocess.PIPE,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+            )
+
+            _, _ = p.communicate()
+            if p.returncode == 0:
+                return self.process_directory(dirpath, duplicates, password)
+
             return []
 
-        if temporary:
-            os.unlink(filepath)
+        except Exception:
+            return []
 
-        return self.process_directory(dirpath, duplicates, password)
+        finally:
+            if temporary and os.path.exists(filepath):
+                os.unlink(filepath)
+
+            if os.path.exists(dirpath):
+                shutil.rmtree(dirpath)
+
+    def get_metadata(self):
+        ret = []
+        content = self.f.contents
+        if not content:
+            return ret
+
+        if b"BEGIN PGP PUBLIC KEY BLOCK" in content:
+            ret.append("public_key")
+        elif b"BEGIN PGP PRIVATE KEY BLOCK" in content:
+            ret.append("private_key")
+        elif b"BEGIN PGP MESSAGE" in content:
+            ret.append("encrypted_message")
+        elif b"BEGIN PGP SIGNATURE" in content:
+            ret.append("signature")
+        elif content[0] & 0x80:
+            # Binary analysis
+            tag = content[0]
+            if tag & 0x40:  # New format
+                tag_type = tag & 0x3F
+            else:  # Old format
+                tag_type = (tag >> 2) & 0xF
+
+            if tag_type in (6, 14):
+                ret.append("public_key")
+            elif tag_type == 5:
+                ret.append("private_key")
+            elif tag_type in (1, 18):
+                ret.append("encrypted_message")
+
+        return ret
